Split out non-relevant-with-defaults code
